# server/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ukr-Book Auth API",
    "version": "1.0.0",
    "description": "API авторизації та технічних endpoint-ів проєкту Ukr-Book."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Операції авторизації користувача"
    },
    {
      "name": "System",
      "description": "Системні endpoint-и"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "login",
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string",
            "example": "user123"
          },
          "firstName": {
            "type": "string",
            "example": "Dmytro"
          },
          "lastName": {
            "type": "string",
            "example": "Medvid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@mail.com"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string",
            "example": "user123"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd123"
          }
        }
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1740000000000"
          },
          "login": {
            "type": "string",
            "example": "user123"
          },
          "firstName": {
            "type": "string",
            "example": "Dmytro"
          },
          "lastName": {
            "type": "string",
            "example": "Medvid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@mail.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          }
        }
      },
      "TokenPayloadUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1740000000000"
          },
          "login": {
            "type": "string",
            "example": "user123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@mail.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "iat": {
            "type": "integer",
            "example": 1710000000
          },
          "exp": {
            "type": "integer",
            "example": 1710600000
          }
        }
      },
      "AuthSuccessResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/TokenPayloadUser"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Невірний логін або пароль."
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Перевірка доступності сервера",
        "description": "Технічний endpoint для перевірки того, що backend працює.",
        "responses": {
          "200": {
            "description": "Сервер доступний",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "ok": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Реєстрація нового користувача",
        "description": "Створює нового користувача, хешує пароль і повертає JWT-токен та дані користувача.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "login": "user123",
                    "firstName": "Dmytro",
                    "lastName": "Medvid",
                    "email": "user@mail.com",
                    "password": "Passw0rd123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Користувача успішно зареєстровано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "user": {
                        "id": "1740000000000",
                        "login": "user123",
                        "firstName": "Dmytro",
                        "lastName": "Medvid",
                        "email": "user@mail.com",
                        "role": "user"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Помилка валідації вхідних даних",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "message": "Будь ласка, заповніть усі обов’язкові поля."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт через зайнятий логін або email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "duplicateLogin": {
                    "value": {
                      "message": "Такий логін вже зайнятий."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вхід користувача в систему",
        "description": "Перевіряє логін або email та пароль, після чого повертає JWT-токен і дані користувача.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "login": "user123",
                    "password": "Passw0rd123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успішна авторизація",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "user": {
                        "id": "1740000000000",
                        "login": "user123",
                        "firstName": "Dmytro",
                        "lastName": "Medvid",
                        "email": "user@mail.com",
                        "role": "user"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Не передано логін або пароль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingData": {
                    "value": {
                      "message": "Будь ласка, введіть логін і пароль."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Невірний логін або пароль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidCredentials": {
                    "value": {
                      "message": "Невірний логін або пароль."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Отримання даних поточного користувача",
        "description": "Повертає payload користувача з валідного JWT-токена.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Дані поточного авторизованого користувача",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "user": {
                        "id": "1740000000000",
                        "login": "user123",
                        "email": "user@mail.com",
                        "role": "user",
                        "iat": 1710000000,
                        "exp": 1710600000
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Токен відсутній або невалідний",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingToken": {
                    "value": {
                      "message": "No token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}